api/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.0.0,<3.0.0
httpx>=0.27.0,<1.0.0
paho-mqtt>=2.0.0,<3.0.0
pyyaml>=6.0,<7.0
python-multipart>=0.0.9
dnspython>=2.6.0,<3.0.0
cachetools>=5.3.0,<6.0.0
cryptography>=42.0.0,<43.0.0
